% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ApplyPeakBoundary}
\alias{ApplyPeakBoundary}
\title{Extract the peak from a chromatogram by restricting the chromatogram to peak boundaries.}
\usage{
ApplyPeakBoundary(peak, boundary, ...)
}
\arguments{
\item{peak}{A peak group of class peakObj}

\item{boundary}{a numeric vector of size two with min start time and max end time of the peak (peak boundaries)}
}
\value{
A peak group of class peakObj with time and intensity vectors that are within the boundaries of the peak
}
\description{
The function takes a peak group of class peakObj and the peak boundary as input. The input peak group is usually a chromatogram with a time range that is wider than the peak of interest. This function extracts peak of interest from the chromatogram by limiting the time and signal intensities of the peak to a range that is within the provided peak boundaries.
}
\examples{

chrom <- data.CSF$data$ChromGroup[[1]]
PlotChromPeak(chrom)
peak <- ApplyPeakBoundary(chrom,c(data.CSF$data$MinStartTime[[1]],data.CSF$data$MaxEndTime[[1]]))
PlotChromPeak(peak)
}
