% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{ApplyQCModel}
\alias{ApplyQCModel}
\title{Apply a binary predictive peak QC model to flag peaks that suffer from poor chromatography or interference.}
\usage{
ApplyQCModel(data.feature, model, response.var = c("Status"),
  description.columns = c("Notes"), flag.prob.threshold = 0.5,
  standard.intensity.threshold = NULL, type = NULL, ...)
}
\arguments{
\item{data.feature}{A dataframe that contains peak identifiers (File,FileName,PeptideModifiedSequence,FragmentIon,IsotopeLabelType,PrecursorCharge and ProductCharge) as well as  QC metrics calcualted for each transition pair. This dataframe is the output of ExtractFeatures function (output$features).}

\item{model}{The predictive model of peak QC. The model is the output of TrainQCModel.}

\item{response.var}{If the input dataframe contains columns corresponding to response variables, it should be indicated here. it should be indicated here. Response and description columns as well as identifier columns will be removed from the data before applying the model.}

\item{description.columns}{If the input dataframe contains columns corresponding to description variables (Such as Notes), it should be indicated here. Response and description columns as well identifier columns will be removed from the data before applying the model.}

\item{flag.prob.threshold}{A numeric value between 0 and 1 which determines the cut-off threshold for assigning classes to each peak based on corresponding class probabilitis. By default, the caret package uses a probability threshold of 0.5. This parameter can be used to overrise the default probability threshold.}

\item{standard.intensity.threshold}{This parameter can be used to set an intensity threshold to identify and flag transitions where the spiked-in standard is too low. If the numerical value of desired intensity threshold is provided, it is used to flag any transition whose standard signal intensity is below this threshold. For such transitions, this will override the model output.}

\item{type}{If type = "prob", the function will return class probabilities for the binary classification. This feature can be used only if the model supports classification probabilities e.g. logistic regression and random forest.}
}
\value{
A dataframe of the predicted response (final class and/or class probabilities)  appended to the input data.feature.
}
\description{
The function takes the feature dataframe (output of ExtractFeatures output$features) and the trained predictive QC model (output of TrainQCModel) and applies the model to the input feature data using the function provided in the caret package.
}
\examples{

response.data <- ApplyQCModel(data.set.CSF$feature.data,
                              model.rrf.CSF,
                              response.var = c("Status"),
                              description.columns = c("Notes"),
                              flag.prob.threshold = 0.5,
                              type = "prob")
}
