% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{BuildPeakGroup}
\alias{BuildPeakGroup}
\title{Convert a dataframe with Times and Intensities columns into a peakObj object}
\usage{
BuildPeakGroup(df, ...)
}
\arguments{
\item{df}{the dataframe with Times and Intensities columns for creating the peakObj object}
}
\value{
A peak group of class peakObj
}
\description{
This is a helper function to format the output of Skyline into peakObj that are compatible with TargetedMSQC functions.
}
