% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{CalculateFWHM}
\alias{CalculateFWHM}
\title{Compute FWHM and ratio of FWHM to peak width at base for transition peaks in a peak group of class peakObj.}
\usage{
CalculateFWHM(peak, ...)
}
\arguments{
\item{peak}{A peak group of class peakObj}
}
\value{
A list with the following objects:
              r.fwhm: A numeric vector of FWHM for each transition in the peak group.
              peak.fwhm: Mean of the r.fwhm vector.
              r.fwhm2base: A numeric vector of FWHM to peak width at base ratio for each transition in the peak group.
              peak.fwhm2base: Mean of the r.fwhm2base vector.
}
\description{
The function takes a peak group of class peakObj as input and calculates the full-width at half-max (FWHM) and the FWHM to peak width at the base ratios for the transition peaks in the peak group. See \url{https://en.wikipedia.org/wiki/Full_width_at_half_maximum}
}
\examples{

peak <- data.CSF$data$PeakGroup[[1]]
peak.group.fwhm <- CalculateFWHM(peak)
}
