% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{CalculatePeakSymmetry}
\alias{CalculatePeakSymmetry}
\title{Compute symmetry scores for transition peaks in a peak group of class peakObj.}
\usage{
CalculatePeakSymmetry(peak, ...)
}
\arguments{
\item{peak}{A peak group of class peakObj}
}
\value{
A list with the following objects:
              r.symmetry: A numeric vector of symmetry scores for each transition in the peak group.
              peak.symmetry: Mean of the r.symmetry vector. This score represents the overall symmetry of all the transitions in the peak group.
}
\description{
The function takes a peak group of class peakObj as input and calculates the symmetry score for the transition peaks in the peak group. The symmetry score is defined as the Pearson correlation coefficient between the left and right halves of the peak intensity vector divided at the timepoint in the middle. For high quality peaks, the symmetry score is expected to be close to 1.
}
\examples{

peak <- data.CSF$data$PeakGroup[[1]]
peak.group.symmetry <- CalculatePeakSymmetry(peak)
}
