% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{CleanUpChromatograms}
\alias{CleanUpChromatograms}
\title{Convert the exported skyline chromatogram and peak boundary reports to a dataframe format compatible with TargetedMSQC functions.}
\usage{
CleanUpChromatograms(chromatogram.path = NULL, peak.boundary.path = NULL,
  labkey.url.base = NULL, labkey.url.path = NULL, parallel = FALSE,
  labkey = FALSE, endogenous.label = "light", standard.label = "heavy",
  iRT.list = c("LGGNETQVR", "AGGSSEPVTGLADK", "VEATFGVDESANK", "YILAGVESNK",
  "TPVISGGPYYER", "TPVITGAPYYER", "GDLDAASYYAPVR", "DAVTPADFSEWSK",
  "TGFIIDPGGVIR", "GTFIIDPAAIVR", "FLLQFGAQGSPLFK", "LGGNEQVTR",
  "GAGSSEPVTGLDAK", "VEATFGVDESNAK", "YILAGVENSK", "TPVISGGPYEYR",
  "TPVITGAPYEYR", "DGLDAASYYAPVR", "ADVTPADFSEWSK", "GTFIIDPGGVIR",
  "GTFIIDPAAVIR", "LFLQFGAQGSPFLK"), ...)
}
\arguments{
\item{chromatogram.path}{Path to the directory containing the .tsv files of the peak chromatograms. For each Skyline document, this file is exported from Skyline through File > Export > Chromatograms. Here, check runs of interest and include Precursors, Products, Base Peaks and TICs. Each chromatogram .tsv file corresponds to a single Skyline document, which may contain any number of runs. Multiple chromatogram files, corresponding to multiple Skyline documents can be copied into the chromatogram.path directory. For each chromatogram file in this folder, there should be a peak boundary file with an identical name in peak.boundary.path directory.}

\item{peak.boundary.path}{Path to the directory containing the .csv files of the peak boundaries. For each Skyline document, this file is exported from Skyline through File > Export > Report. Here, select Peak Boundaries. Each peak boundary .csv file corresponds to a single Skyline document, which may contain any number of runs. Multiple peak boundary files, corresponding to multiple Skyline documents can be copied into the peak.boundary.path directory. For each peak boundary file in this folder, there should be a peak chromatogram file with an identical name in chromatogram.path directory.}

\item{labkey.url.base}{URL of the labkey server, if data is being imported from labkey (labkey = TRUE). This feature has not been implemented yet.}

\item{labkey.url.path}{Path to the directory containing the Skyline documents on the labkey server, if data is being imported from labkey (labkey = TRUE). This feature has not been implemented yet.}

\item{parallel}{Logical parameter to determine whether the function should run in parallel. This feature has not been implemented yet. Set parallel = FALSE.}

\item{labkey}{Logical parameter to indicate whether data will be imported from labkey. This feature has not been implemented yet. Set labkey = FALSE.}

\item{endogenous.label}{Label of the endogenous analyte: default is "light"}

\item{standard.label}{Label of the spiked-in isotopically labeled analyte: default is "heavy"}

\item{iRT.list}{List of iRT standards used in the experiment. These peptides will be removed from the training set.}
}
\value{
A list with the following objects:
              data: A dataframe that contains all rows of the input data with assigned peak boundaries. This dataframe also includes columns for peakObj objects created for each peak group.
              removed: A dataframe that contains the rows with missing peak boundary values, too few transitions and missing isotope pairs. These rows are removed from downstream feature extraction and QC analysis.
}
\description{
The function takes the directories that contain chromatogram and peak boundary files of Skyline documents and merges them. The rows with NA values for peak boundaries, rows corresponding to peptides with fewer than 3 transitions and transitions missing a standard isotope pair are separated into a dataframe that is returned by the function in output$removed and excluded from subsequence QC steps. The remainder of the rows are returned by the function in output$data for subsequence analysis. Transition peaks of each peptide in each run are formed into peak groups of custom class peakObj and stored in the ChromGroup column. The ApplyPeakBoundary function is applied to chromatograms to restrict them to the provided peak boundaries and the results are stored in the PeakGroup column. Additionally, rows are added for the "sum" transition, which are is the sum of peptides transitions with the same isotope label in each sample.
}
\examples{

extdata.path <- system.file("extdata",package = "TargetedMSQC")
project.folder.name <- "CSF_Panel"
project.path <- file.path(extdata.path,project.folder.name)
chromatogram.path <- file.path(project.path,"Chromatograms")
peak.boundary.path <- file.path(project.path,"Peak_boundary")
data <- CleanUpChromatograms(chromatogram.path = chromatogram.path,
                            peak.boundary.path = peak.boundary.path,
                            endogenous.label = "light",
                            standard.label = "heavy",
                            iRT.list = c("LGGNETQVR","AGGSSEPVTGLADK","VEATFGVDESANK","YILAGVESNK","TPVISGGPYYER","TPVITGAPYYER","GDLDAASYYAPVR","DAVTPADFSEWSK","TGFIIDPGGVIR","GTFIIDPAAIVR","FLLQFGAQGSPLFK"))
}
