% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{ExtractFeatures}
\alias{ExtractFeatures}
\title{Calculate an ensemble of QC features/metrics for a dataset that has been cleaned up by CleanUpChromatograms.}
\usage{
ExtractFeatures(data, parallel = FALSE, blanks = NA,
  intensity.threshold = 1e+05, endogenous.label = "light",
  standard.label = "heavy", export.features = FALSE, feature.path = "",
  ...)
}
\arguments{
\item{data}{A dataframe that contains identifier columns and peak groups of class peakObj for each transition peak. This dataframe is the output of CleanUpChromatograms (output$data).}

\item{parallel}{Logical parameter to determine whether the function should run in parallel. This feature has not been implemented yet. Set parallel = FALSE.}

\item{blanks}{A dataframe with two columns of File and FileName. The FileName column contains the names of blank runs for each Skyline File in the File column. The values in File and FileName should be consistent with corresponding columns in input data. Example: If Skyline document "SkylineFile1" includes three blank runs of "Blank1", "Blank2" and "Blank3", blanks <- data.frame(File = "SkylineFile1", FileName = c("Blank1","Blank2","Blank3")). The function uses the blank runs to estimate the peak intensity at LLOQ for each pepide and transition. If blanks = NA, the function uses the intensity.threshold by default as approximation of intensity at LLOQ for all peaks. This threshold can be adjusted.}

\item{intensity.threshold}{The threshold that the function uses as an approximation of the intensity at LLOQ for all peptides. If blank samples are not provided to estimate the intensity at LLOQ, this value is used as an approximation to determine transition peaks that are below limit of quantitation.}

\item{endogenous.label}{Label of the endogenous analyte: default is "light"}

\item{standard.label}{Label of the spiked-in isotopically labeled analyte: default is "heavy"}

\item{export.features}{Logical parameter to indicate whether calculated features should be exported as a .csv file.}

\item{feature.path}{Path to the directory, where the features should be saved if export.features = TRUE.}
}
\value{
A list with the following objects:
              features: A dataframe with columns that contain the calculated QC features for each transition pair in the input data.
}
\description{
The function takes the output of CleanUpChromatograms as input and for each transition pair of each peptide, calculates a number of QC metrics that are used as features for training a predictive peak QC model.
}
\examples{

data.features <- ExtractFeatures(data = data.CSF$data,
                                export.features = FALSE,
                                intensity.threshold = 1000)
}
