% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{iRTList}
\alias{iRTList}
\title{This file contains all the helper functions required for the qc process
Return a vector of all the iRT standard used for the experiment. This is}
\usage{
iRTList()
}
\description{
This file contains all the helper functions required for the qc process
Return a vector of all the iRT standard used for the experiment. This is
}
\examples{

standards <- iRTList()
standards
}
