% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{methods}
\name{is.na,peakObj-method}
\alias{is.na,peakObj-method}
\title{Default S3 is.na method for peakObj
peakObj object is declared na if the time or sig portion of it is of length 0}
\usage{
\S4method{is.na}{peakObj}(x)
}
\description{
Default S3 is.na method for peakObj
peakObj object is declared na if the time or sig portion of it is of length 0
}
