% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{MakeDataSet}
\alias{MakeDataSet}
\title{Create the dataset used as input to the TrainQCModel function.}
\usage{
MakeDataSet(feature.data = NULL, feature.path = NULL,
  training.path = NULL, ...)
}
\arguments{
\item{feature.data}{A dataframe that contains peak identifiers (File,FileName,PeptideModifiedSequence,FragmentIon,IsotopeLabelType,PrecursorCharge and ProductCharge) as well as  QC metrics calcualted for each transition pair. This dataframe is the output of ExtractFeatures function (output$features).}

\item{feature.path}{Alternative to providing the feature.data, a path to the directory that contains the features.csv file exported by ExtractFeatures function can be provided. This file can be generated by by setting export.features = TRUE and specifying the feature.path in ExtractFeatures function. If both feature.data and feature.path are provided, feature.path is ignored.}

\item{training.path}{Path to the directory containing the annotated template file. The template file is generated using the MakeTemplate function should be manually annotated by an expert analyst and saved as a separate .csv file. The path to this annotated file should be provided. Please note that the directory should contain only annotated .csv files that are meant to be in the study.}
}
\value{
A list with the following objects:
              data.merged: A dataframe that is product of merging and cleaning up feature.data and training.data. This data can be used by the TrainQCModel for training a predictive peak QC model.
              feature.data: The input feature.data
              training.data: The annotated input training.data
}
\description{
The function takes the feature dataframe (containing the calculated ensemble of QC features) and the annotated training dataframe and merges them to create the input of the TrainQCModel function.
}
\examples{

extdata.path <- system.file("extdata",package = "TargetedMSQC")
project.folder.name <- "CSF_Panel"
project.path <- file.path(extdata.path,project.folder.name)
training.path <- file.path(project.path,"Training")
data.set <- MakeDataSet(feature.data = data.features.CSF$features,training.path = training.path)
feature.path <- file.path(project.path,"Features")
data.set <- MakeDataSet(feature.path = feature.path,training.path = training.path)
}
