% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{MakeTemplate}
\alias{MakeTemplate}
\title{Create a training set template that can be shared with an analyst for manual annotation.}
\usage{
MakeTemplate(chromatogram.path, template.path, training.filename.list = "all",
  endogenous.label = "light", standard.label = "heavy",
  iRT.list = c("LGGNETQVR", "AGGSSEPVTGLADK", "VEATFGVDESANK", "YILAGVESNK",
  "TPVISGGPYYER", "TPVITGAPYYER", "GDLDAASYYAPVR", "DAVTPADFSEWSK",
  "TGFIIDPGGVIR", "GTFIIDPAAIVR", "FLLQFGAQGSPLFK", "LGGNEQVTR",
  "GAGSSEPVTGLDAK", "VEATFGVDESNAK", "YILAGVENSK", "TPVISGGPYEYR",
  "TPVITGAPYEYR", "DGLDAASYYAPVR", "ADVTPADFSEWSK", "GTFIIDPGGVIR",
  "GTFIIDPAAVIR", "LFLQFGAQGSPFLK"), ...)
}
\arguments{
\item{chromatogram.path}{Path to the directory containing the .tsv files of the peak chromatograms. For each Skyline document, this file is exported from Skyline through File > Export > Chromatograms. Here, check runs of interest and include Precursors, Products, Base Peaks and TICs. Each chromatogram .tsv file corresponds to a single Skyline document, which may contain any number of runs. Multiple chromatogram files, corresponding to multiple Skyline documents can be copied into the chromatogram.path directory. For each chromatogram file in this folder, there should be a peak boundary file with an identical name in peak.boundary.path directory.}

\item{template.path}{Path to the directory, where the template file will be saved.}

\item{training.filename.list}{List of the runs that are going to be used for training. if set to defaul of "all", all the runs in the chromatogram file will be used for training.}

\item{endogenous.label}{Label of the endogenous analyte: default is "light"}

\item{standard.label}{Label of the spiked-in isotopically labeled analyte: default is "heavy"}

\item{iRT.list}{List of iRT standards used in the experiment. These peptides will be removed from the training set.}
}
\value{
Saves the template .csv file in template.path
}
\description{
The function takes the directory that contains chromatogram of Skyline documents and creates a template for manual annotation of the peaks. The template is saved as a .csv file in the directory specified by template.path. This templae should be shared with an analyst for manual annotation and the "Status" column should be filled with "flag" or "ok" for low and high quality peaks, respectively. Any comments about the manual annotation can be entered in the "Notes" column.
}
\examples{

extdata.path <- system.file("extdata",package = "TargetedMSQC")
project.folder.name <- "CSF_Panel"
project.path <- file.path(extdata.path,project.folder.name)
chromatogram.path <- file.path(project.path,"Chromatograms")
template.path <- file.path(project.path,"Templates")
MakeTemplate(chromatogram.path = chromatogram.path,
              template.path = template.path,
              endogenous.label = "light",standard.label = "heavy")
}
