% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{class}
\name{peakObj-class}
\alias{peakObj-class}
\alias{peakObj}
\title{A custom class for storing single peaks or multiple peaks that belong to the same peptide and sample in a single object}
\description{
A custom class for storing single peaks or multiple peaks that belong to the same peptide and sample in a single object
}
\section{Slots}{

\describe{
\item{\code{time}}{A numeric vector of the peak timepoints}

\item{\code{sig}}{A numeric vector or dataframe of the peak signal intensities. If this is a data}

\item{\code{area}}{A numeric vector of the AUC of peaks in the sig slot}
}}

\examples{

peak <- data.CSF$data$PeakGroup[[1]]
peak@time
peak@sig
peak@area
}
