% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_chrom}
\alias{plot_chrom}
\title{Plot Chromatograms}
\usage{
plot_chrom(
  peak,
  font.size = 14,
  transition.list = NA,
  label.list = NA,
  split.label = TRUE
)
}
\description{
Plot Chromatograms
}
