% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PlotChromPeak}
\alias{PlotChromPeak}
\title{Plot the chromatographic peaks of class peakObj}
\usage{
PlotChromPeak(peak, plottype = c("ggplot", "plotly"), split.label = TRUE,
  endogenous.label = "light", standard.label = "heavy",
  transition.list = NULL, label.list = NULL, font.size = 20, ...)
}
\arguments{
\item{peak}{A peak group of class peakObj}

\item{plottype}{This parameter can take values of "ggplot" for static and "plotly" for interactive plots}

\item{split.label}{A logical parameter. If split.label = TRUE, the peaks corresponding to different isotope label types will be split into two panels}

\item{endogenous.label}{Label of the endogenous analyte: default is "light"}

\item{standard.label}{Label of the spiked-in isotopically labeled analyte: default is "heavy"}

\item{transition.list}{List of transitions to be plotted. If transition.list = NULL, all transitions will be plotted. Example: transition.list = c("b3","y4")}

\item{label.list}{List of isotope label types to be plotted. If label.list = NULL, all isotope label types will be plotted. Example: label.list = c("light")}

\item{font.size}{Font size of text in the plot}
}
\value{
A plot object of class ggplot or plotly
}
\description{
The function takes a peak group of class peakObj as input and plots its chromatogram. It can generate a static (ggplot) or interactive (plotly) plot.
}
\examples{

peak <- data.CSF$data$PeakGroup[[1]]
PlotChromPeak(peak,transition.list = c("b5","y6"), label.list = NULL,split.label = FALSE)
PlotChromPeak(peak,split.label = TRUE)
}
