% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetedMSQC.R
\docType{package}
\name{TargetedMSQC}
\alias{TargetedMSQC}
\alias{TargetedMSQC-package}
\title{TargetedMSQC: a tool for quality assessment and interference detection in targeted mass spectrometry data using machine learning}
\description{
TargetedMSQC provides a semi-automated workflow for quality control (QC) of chromatographic peaks in targeted proteomics experiments, with the aim of improving the efficiency and reducing the subjectivity of data QC. The package offers a toolkit to build and apply statistical models for predicting peak qualities in proteomics datasets using supervised learning methods. The package contains functions to calculate an ensemble of >30 well-established and newly introduced peak quality metrics, such as jaggedness, FWHM, modality, shift, coefficient of variation, consistency of transition peak area ratios, etc., to quantify the quality of chromatographic peaks. These quality control metrics calculated in a training dataset of peaks with pre-annotated quality status labels are used as the feature set in supervised learning algorithms to flag peaks with poor chromatography or interference in other targeted proteomics experiments.
}
\seealso{
\url{https://github.com/shadieshghi/TargetedMSQC}
}
\author{
Shadi Eshghi, \email{toghiess@gene.com}
}
\keyword{analysis,}
\keyword{automated}
\keyword{bioinformatics,}
\keyword{control,}
\keyword{detection,}
\keyword{interference}
\keyword{learning}
\keyword{machine}
\keyword{mass}
\keyword{proteomics,}
\keyword{quality}
\keyword{quantitative,}
\keyword{spectrometry,}
\keyword{targeted}
