% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{CalculateMaxBoundaryIntensity}
\alias{CalculateMaxBoundaryIntensity}
\title{Compute maximum intensity at the peak boundary of each transition peak in a
peak group object.}
\usage{
CalculateMaxBoundaryIntensity(peak, ...)
}
\arguments{
\item{peak}{A peak group object}
}
\value{
A numeric vector of maximum intensity at peak boundary for individual
transition peaks in the peak group
}
\description{
The function takes a peak group object as input and returns a numeric vector
of maximum intensity at peak boundary for individual transition peaks in the
peak group.
}
\examples{

peak <- data.CSF$data$PeakGroup[[1]]
transition.max.at.boundary.intensities <- CalculateMaxBoundaryIntensity(peak)
}
